Fixed bug in saving individual results
Fixed a bug which saved MANY individual datastructs, which was annoying and unnecessary. Individual curve analysis now results in the same as the average analysis (a single mat file with the datastruct), which just has results in rows 1-N (for N curves) instead of N+M (with M velocities present).